Global_AI_Travel _Guide/requirements.txt
openai
python-dotenv
geopy
Pillow
requests
wikipedia
tkintermapview
